+++ src/main/java/me/matrix89/markpages/data/repository/UserRepository.java
@@ -3,6 +3,6 @@ package me.matrix89.markpages.data.repository;
 import me.matrix89.markpages.data.model.UserModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserModel, Integer> {
+public interface UserRepository extends CrudRepository<UserModel, Long> {
     UserModel getByUsername(String username);
 }
